+++ src/builder/node.js
@@ -202,6 +202,9 @@ class BuilderNode {
 		let params = builder.getParentCallbackParameters();
 		// Bind builder to this, so that this.xxx can be used for processes
 		let node = arg.bind(builder)(...params);
+		if (builder.supports('array_create') && Array.isArray(node)) {
+			node = builder.array_create(node);
+		}
 		if (node instanceof BuilderNode) {
 			node.result = true;
 			return builder.toJSON();